move yaml func to file utils
